Changed back to Integer casts of criteria.uniqueResult() (Hibernate version 3.3.2.GA returns Integer)
git-svn-id: svn://localhost/ARK@6467 35ee9b83-dd2c-47f4-99a2-c706670c48ce

+++ ark-common/src/main/java/au/org/theark/core/dao/ArkAuthorisationDao.java
@@ -1064,7 +1064,7 @@ public class ArkAuthorisationDao<T> extends HibernateSessionDao implements IArkA
 		Criteria criteria = getSession().createCriteria(ArkUserRole.class);
 		criteria.add(Restrictions.eq("arkUser", arkUser));
 		criteria.setProjection(Projections.rowCount());
-		Long totalCount = (Long) criteria.uniqueResult();
+		Integer totalCount = (Integer) criteria.uniqueResult();
 		if (totalCount > 0) {
 			flag = true;
 		}
